.travis.yml: Test no_std platform, WASM, all feature combos, --release
This uses a CI config more like the one at:

https://github.com/RustCrypt/signatures

It runs all tests with the `--release` flag to ensure that crates aren't
broken by release-dependent functionality after being tested in debug
mode (we can add back debug mode testing if it's helpful, but it doesn't
look like we have anything presently gated on debug builds). I now see
@newpavlov also requested it on #20.

It also tests on a `no_std` platform (thumbv7) which @newpavlov
requested in #18.

Finally, it tests with `--all-features`. This should uncover some
lingering breakages (it won't in this commit, but it appears some of the
conditional compilation is incorrectly gated and I intend to fix it in a
follow-up commit).