impl SyncStreamCipher for Salsa20 and ChaCha20
Previously they only impl'd `StreamCipher`

Additionally this fixes bugs in the previous `SyncStreamCipherSeek`
impls on both ciphers by better encapsulating the block index and offset
management in the `SalsaFamilyCipher` type. This allows the crate
examples to use this impl in the documented examples, which is
consistent with the other crates.