chacha20: Extract and encapsulate Cipher type
This commit primarily renames the previous `ChaChaState` type and
factors it under `chacha20::cipher::Cipher`.

Additionally, it gets rid of a number of unnecessary things from the
previous implementation:

- `ArrayLength` generic parameter
- `Deref`/`DerefMut` impls
- Macros to define them

Instead, `ChaCha20` and `ChaCha20Legacy` are now proper newtypes of
`Cipher`, sharing a completely identical core cipher type, but also
fully encapsulating it so it's no longer part of the public-facing API
whatsoever.

To do that, each of `ChaCha20` and `ChaCha20Legacy` each impl
`NewStreamCipher`, `SyncStreamCipher`, and `SyncStreamCipherSeek`
directly. This is actually a good and important thing for
`NewStreamCipher` in particular, since that is where the differences
between the two of them actually exist.

Overall I'd say this is a lot cleaner, and hides the
`ChaChaState`/`Cipher` type from the public-facing API.