chacha20: Buffering bugfixes and AVX2 backend refactoring
Previously when looping through input data, if there was enough data to
avoid buffering it was obliterating the plaintext with the keystream
rather than XORing it.

This was addressed by adding an `apply_keystream` method to each backend
for use in this situation, which avoids the intermediate buffer.

Additionally this commit refactors the AVX2 backend to look more like
the SSE2 backend.