chacha20: Parallelize AVX2 backend
The AVX2 backend was previously computing two ChaCha blocks in parallel,
then throwing one away.

This updates the implementation to always compute two blocks in parallel
when the AVX2 backend is enabled, resulting in a ~2X speedup.

Unfortunately for `cipher.rs`, originally adapted from the `ctr` crate,
I deleted the original parallel computation code, and in lieu of that
the implementation diverges from what was originally in `ctr`. See here
for a reference:

https://github.com/RustCrypto/stream-ciphers/blob/e8f02d1/ctr/src/lib.rs#L73

Ideally we can come up with some generic counter management and
buffering abstraction in the `ctr` crate which works in all cases.