salsa20: replace `salsa20-core` with `ctr`-derived buffering; MSRV 1.34+
This is pretty much the same PR as #81 was for `chacha20`.

The previously used buffering logic in the `salsa20-core` crate is
complex, hard-to-audit, and slow:

https://github.com/RustCrypto/stream-ciphers/blob/9d51f83/salsa20-core/src/lib.rs#L93

The main reason is because it tries to handle too many things at once,
namely:

- serializing the Salsa20 block function output as little endian bytes
- implementing the stream-cipher APIs

This commit gets rid of `salsa20-core` as a dependency, replacing it
with code derived from the `ctr` crate, but specialized to the Salsa20
use case.

Ideally this code could eventually be unified with the extremely similar
code in the `ctr` crate and `chacha20` crates, as much of it seems
reusable.

For now though, I think it's probably better to keep the code in
`chacha20` separate from this code: the previous API was a major
impediment in SIMD optimizations, and being able to refactor both the
buffering logic and the various block function implementations
internally to a single crate is much easier than having to touch
(and release) 3 different crates for such a change.

This crate doesn't have a SIMD backend, and there's no plans to add one,
so using the simplest buffering logic possible probably makes sense.

As this was the last crate with a code dependency on `salsa20-core`,
this commit also removes it, as it's no longer needed.