chacha20: replace macros with Rounds trait + generics
Previously macros were used to generate the ChaCha8, ChaCha12, and
ChaCha20 variants of the ChaCha family.

This commit replaces them with a `Rounds` trait, making the `Cipher`
type generic over `Rounds`, and replacing the `ChaCha8`, `ChaCha12`,
and `ChaCha20` wrappers with type aliases.

In addition to everything else, this makes the number of rounds type
safe in that they can only be `R8`, `R12`, or `R20`, eliminating runtime
checks around the number of rounds, and also ensuring the number is a
constant value known at compile time, which should improve
optimizations.

The random number generator types are still generated by macros,
unfortunately, but these can be fixed up in a different pass.